Update autotune.md (Capped Autotune in subhead)
There's an argument to be made that all the proper names of the features in iAPS, like Autotune, Autosens, Dynamic ISF, etc. should be capped consistently so it's clear these are features and not an action. Not to belabor this small point but it appears on the OpenAPS site they cap these in page titles, subheads, etc, but not necessarily in the body copy. Small changes like that can help new users--like, hey, this is a feature setting I need to adjust not a general concept. Helps people find them in the menus too.