Update carbratios.md (typo, suggested edit)
Subjunctive tense, was to were. But I'm not clear on what the last sentence on the page means. 

Does it mean once autotune stabilizes, you can use the Autotune ICR setting instead of the one(s) you have in the Carb Ratios settings under Schedule? 

Not clear on what it means to use Autotune vs. "being limited by a safety limiter." I'd try to explain that simply or cut.