Update smbsettings.md
This is a suggested edit to clarify the purpose of the "Enable SMB With High BG" setting, and make the connection clearer to the "… When Blood Glucose is Over (mg/dl)" setting. The documentation, as written, doesn't make the connection clear. There's a line between them in the interface, which could suggest they aren't connected and the elipsis before the next setting could mean it's connected to all the SMB settings. Discussed on Discord Jon Björn said: "Yes, I’ve thought many times about removing this new setting." The feature itself can override others, like scheduling when SMBs are off in a profile.  It also seems in conflict with the Threshold Setting, which I think it what most users would be attempting to do--not get insulin below a certain BG level. Jon also said: "The 'Enable SMB with High BG' is causing much confusion. I would suggest to all to not use this setting.
Because it overrides also the temp target intention
If we should keep this setting in iAPS we need to add a warning and a detailed description." [I may not have formatted the suggested change right, regrets]